vendor: Incluse some gapps from BeansGapps
Thanks to BeansGapps dev for all the work on making gapps work. The only thing i've done is adapt all the stuff for use the keyboard and phone app. And i had to change the name because the previous name was too STOCK-ROM-like. And opengapps will delete it just for that.
Shishu-ify: Adding the new stuff to main makefile.