WORKAROUND: Use beansgapps in other folder.
I think at the time of building, it reads the first android.mk, then everything else is over (i mean, if you have a Android.mk on a subfolder). That was breaking beansgapps from building and my attempts to fix it breaks and FCs the keyboard and phone. So why i don't leave it in a separate folder and call the makefile like the old times? but this time i think it'll work fine because it isn't any other Android.mk before the beansgapps folder.